requirements.txt
tensorflow 
gradio
numpy
pillow